app/requirement.txt
absl-py
aiodns
aiohappyeyeballs
aiohttp
aiosignal
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
astunparse
async-lru
async-timeout
attrs
babel
beautifulsoup4
bleach
Brotli
cached-property
cachetools
certifi
cffi
charset-normalizer
click
colorama
comm
contourpy
cryptography
cycler
debugpy
decorator
defusedxml
exceptiongroup
executing
fastjsonschema
filelock
flatbuffers
fonttools
fqdn
frozenlist
gast
gmpy2
google-auth
google-auth-oauthlib==0.4.1
google-pasta
grpcio
h11
h2
h5py
hpack
httpcore
httpx
hyperframe
idna
importlib_metadata
ipykernel
ipython
isoduration
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
keras
kiwisolver
lark
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
ml_dtypes
mpi4py
mpmath
multidict
munkres==1.1.4
namex
narwhals
nbclient
nbconvert
nbformat
nest_asyncio
networkx
notebook_shim
numpy
oauthlib
opencv-python>=4.10.0
opencv-python-headless>=4.10.0
opt_einsum
optree
overrides
packaging
pandas
pandocfilters
parso
patsy
pexpect
pickleshare
Pillow
platformdirs
plotly
prometheus_client
prompt_toolkit
propcache
protobuf==5.29.3
psutil
ptyprocess==0.7.0
pure_eval
pyasn1
pyasn1_modules
pycares
pycparser
Pygments
pyOpenSSL
pyparsing
python-dateutil
python-json-logger
pytz
PyYAML
pyzmq
referencing
requests
requests-oauthlib
rfc3339_validator
rfc3986-validator
rfc3987-syntax
rich
rpds-py
rsa
scikit-learn
scipy
seaborn
Send2Trash
six
sniffio
soupsieve
stack_data
statsmodels
sympy
tensorboard==2.20.0
tensorboard-plugin-wit
tensorboard_data_server
tensorflow==2.20.0
tensorflow-estimator
termcolor
terminado
threadpoolctl
tinycss2
tomli
torch==2.5.1
torchaudio>=2.6.0,<2.9.0
torchvision
tornado
tqdm
traitlets
types-python-dateutil
typing_extensions
typing_utils
tzdata
unicodedata2
uri-template
urllib3
uvloop
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
wrapt
yarl
zipp
zstandard
